harness: fold refactor perception into substrate
